+++ Template/Utils/Initialize/Initialize.h
@@ -47,7 +47,7 @@ extern "C" {
 #endif
 extern CarState car; // Declare car state for kinematics
 extern Data current_data; // Declare current data for kinematics
-extern int Digital[8]; // Declare an array for digital sensor readings
+extern uint8_t Digital[8]; // Declare an array for digital sensor readings
 extern char error_message[100]; // Buffer for error messages
 extern double GyroscopeChannelData[10]; // Array to hold gyroscope data
 extern uint32_t tick; // Global tick counter for timing
